More fixes. Many of these are not detected on x86 but on Alpha.

+++ pgserver/widget/box.c
@@ -1,4 +1,4 @@
-/* $Id: box.c,v 1.21 2002/01/14 08:52:38 micahjd Exp $
+/* $Id: box.c,v 1.22 2002/02/02 20:52:52 lonetech Exp $
  *
  * box.c - Generic container for holding a group of widgets. It's sizing and
  *         appearance are defined by the theme.
@@ -29,7 +29,7 @@
 #include <pgserver/common.h>
 #include <pgserver/widget.h>
 
-#define MARGIN_OVERRIDE ((int)self->data)
+#define MARGIN_OVERRIDE ((long)self->data)
 
 void box_resize(struct widget *self) {
    int m;
